% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gft}
\alias{gft}
\title{Graph Fourier Transform}
\usage{
gft(signal, U)
}
\arguments{
\item{signal}{Input signal (vector or matrix)}

\item{U}{Matrix of eigenvectors (dense matrix preferred)}
}
\value{
Transformed signal in the spectral domain (vector or matrix)
}
\description{
Compute the Graph Fourier Transform (GFT) of a signal using Laplacian eigenvectors.
}
