% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hello_sgwt}
\alias{hello_sgwt}
\title{Hello function for SGWT package demonstration}
\usage{
hello_sgwt()
}
\value{
Character string with greeting
}
\description{
Simple hello function to demonstrate package loading
}
\examples{
hello_sgwt()
}
