% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgwt_main.R
\name{print.SGWT}
\alias{print.SGWT}
\title{Print method for SGWT objects}
\usage{
\method{print}{SGWT}(x, ...)
}
\arguments{
\item{x}{SGWT object to print}

\item{...}{Additional arguments passed to print methods}
}
\description{
Print method for SGWT objects
}
